pyproject.toml
[tool.poetry]
name = "wordle_solver"
version = "0.1.0"
description = ""
authors = ["Ilya Isko <ilufkai2000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.27.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.8.0"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
